added route beverages
