+++ app.py
@@ -1,21 +1,22 @@
 """
-This module is the main module for the app.
-It performs all the core features when the app is running.
+Main module for the app.
 """
 import os
 import subprocess
 from flask import Flask, flash, redirect, url_for
 from dotenv import load_dotenv
+from routes.beverages import beverages_bp
 from routes.beverage import beverage_bp
 from routes.locations import locations_bp
 
 app = Flask(__name__)
+app.register_blueprint(beverages_bp, url_prefix='/beverages')
 app.register_blueprint(beverage_bp, url_prefix='/beverage')
 app.register_blueprint(locations_bp, url_prefix='/locations')
 
 @app.route('/')
 def index():
-    return redirect(url_for('beverage_bp.index'))
+    return redirect(url_for('beverages_bp.index'))
 
 with app.app_context():
     # Generate secret key
