+++ templates/beverage_view.html
@@ -3,7 +3,7 @@
 {% block content %}
     <nav aria-label="breadcrumb">
         <ul>
-            <li><a href="{{ url_for('beverage_bp.index') }}">Dashboard</a></li>
+            <li><a href="{{ url_for('beverages_bp.index') }}">Dashboard</a></li>
             <li>{{ beverage["beverageName"] }}{% if beverage["beverageYear"] != "" %} ({{ beverage["beverageYear"] }}){% endif %}</li>
         </ul>
     </nav>
